@@ -1,12 +1,12 @@
 name: flutter_radio_player
 description: Flutter radio plugin handles a single streaming audio preciously.
-version: 1.0.9
+version: 1.1.0
 homepage: https://github.com/Sithira/FlutterRadioPlayer
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.12.0 <2.0.0"
-
+ 
 dependencies:
   flutter:
     sdk: flutter
