@@ -5,8 +5,7 @@ import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
 import android.content.IntentFilter
-import android.media.AudioAttributes
-import android.media.AudioFocusRequest
+import android.graphics.Bitmap
 import android.media.AudioManager
 import android.net.Uri
 import android.os.Binder
@@ -17,10 +16,7 @@ import android.telephony.TelephonyManager
 import androidx.core.app.NotificationCompat
 import androidx.core.app.NotificationManagerCompat
 import androidx.localbroadcastmanager.content.LocalBroadcastManager
-import com.google.android.exoplayer2.C
-import com.google.android.exoplayer2.ExoPlaybackException
-import com.google.android.exoplayer2.Player
-import com.google.android.exoplayer2.SimpleExoPlayer
+import com.google.android.exoplayer2.*
 import com.google.android.exoplayer2.metadata.Metadata
 import com.google.android.exoplayer2.metadata.MetadataOutput
 import com.google.android.exoplayer2.source.MediaSource
@@ -28,6 +24,7 @@ import com.google.android.exoplayer2.source.ProgressiveMediaSource
 import com.google.android.exoplayer2.source.dash.DashMediaSource
 import com.google.android.exoplayer2.source.hls.HlsMediaSource
 import com.google.android.exoplayer2.source.smoothstreaming.SsMediaSource
+import com.google.android.exoplayer2.ui.PlayerNotificationManager
 import com.google.android.exoplayer2.upstream.DefaultDataSourceFactory
 import com.google.android.exoplayer2.util.Util
 import me.sithiramunasinghe.flutter.flutter_radio_player.FlutterRadioPlayerPlugin.Companion.broadcastActionName
@@ -35,7 +32,7 @@ import me.sithiramunasinghe.flutter.flutter_radio_player.FlutterRadioPlayerPlugi
 import me.sithiramunasinghe.flutter.flutter_radio_player.core.enums.PlaybackStatus
 import java.util.logging.Logger
 
-class StreamingCore : Service(), AudioManager.OnAudioFocusChangeListener, MetadataOutput {
+class StreamingCore : Service(), MetadataOutput {
 
     private var logger = Logger.getLogger(StreamingCore::javaClass.name)
 
@@ -56,8 +53,7 @@ class StreamingCore : Service(), AudioManager.OnAudioFocusChangeListener, Metada
     // class instances
     private var telephonyManager: TelephonyManager? = null
     private var audioManager: AudioManager? = null
-    private var audioFocusRequest: AudioFocusRequest? = null
-    private var notification: Notification? = null
+    private var playerManager: PlayerNotificationManager? = null
 
     private var packageIntentName = ""
     private var currentSong = ""
@@ -75,19 +71,25 @@ class StreamingCore : Service(), AudioManager.OnAudioFocusChangeListener, Metada
     fun play() {
         logger.info("playing audio $player ...")
 
-        requestAudioFocus()
+        if (this.playbackStatus == PlaybackStatus.PAUSED) {
+            player.play() // continue playing
+        } else {
+            val dataSourceFactory =
+                DefaultDataSourceFactory(context, Util.getUserAgent(context, initialTitle))
+            val audioSource = buildMediaSource(dataSourceFactory, streamUrl)
 
-        val dataSourceFactory = DefaultDataSourceFactory(context, Util.getUserAgent(context, initialTitle))
-        val audioSource = buildMediaSource(dataSourceFactory, streamUrl)
-
-        player.stop()
-        player.prepare(audioSource)
-        player.playWhenReady = true
+            player.stop()
+            player.setMediaSource(audioSource)
+            player.prepare()
+            player.setForegroundMode(true)
+            player.playWhenReady = true
+            player.play()
+        }
     }
 
     fun pause() {
         logger.info("pausing audio...")
-        removeAudioFocus()
+        player.pause()
         player.playWhenReady = false
     }
 
@@ -99,7 +101,6 @@ class StreamingCore : Service(), AudioManager.OnAudioFocusChangeListener, Metada
 
     fun stop() {
         logger.info("stopping audio $player ...")
-        removeAudioFocus()
         player.stop()
         stopSelf()
     }
@@ -111,7 +112,14 @@ class StreamingCore : Service(), AudioManager.OnAudioFocusChangeListener, Metada
 
     fun setUrl(url: String) {
         logger.info("Set stream URL: $url")
-        streamUrl = url
+        if (isPlaying()) {
+            player.stop(true)
+            player.setMediaItem(MediaItem.fromUri(url))
+            player.prepare()
+            player.playWhenReady = true
+        } else {
+            streamUrl = url
+        }
     }
 
     fun currentSongTitle(): String {
@@ -124,9 +132,9 @@ class StreamingCore : Service(), AudioManager.OnAudioFocusChangeListener, Metada
 
         // get details
         initialTitle = intent!!.getStringExtra("initialTitle")
-        subTitle = intent!!.getStringExtra("subTitle")
-        streamUrl = intent!!.getStringExtra("streamUrl")
-        packageIntentName = intent!!.getStringExtra("packageName")
+        subTitle = intent.getStringExtra("subTitle")
+        streamUrl = intent.getStringExtra("streamUrl")
+        packageIntentName = intent.getStringExtra("packageName")
 
         // init objects
         playbackStatus = PlaybackStatus.IDLE
@@ -136,17 +144,6 @@ class StreamingCore : Service(), AudioManager.OnAudioFocusChangeListener, Metada
         telephonyManager?.listen(phoneStateListener, PhoneStateListener.LISTEN_CALL_STATE)
         audioManager = getSystemService(Context.AUDIO_SERVICE) as AudioManager
         registerReceiver(becomingNoisyReceiver, IntentFilter(AudioManager.ACTION_AUDIO_BECOMING_NOISY))
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            audioFocusRequest = AudioFocusRequest.Builder(AudioManager.AUDIOFOCUS_GAIN)
-                    .setAudioAttributes(
-                            AudioAttributes.Builder()
-                                    .setUsage(AudioAttributes.USAGE_GAME)
-                                    .setContentType(AudioAttributes.CONTENT_TYPE_SPEECH)
-                                    .build()
-                    )
-                    .setAcceptsDelayedFocusGain(false)
-                    .setOnAudioFocusChangeListener(context).build()
-        }
 
         localBroadcastManager = LocalBroadcastManager.getInstance(context)
         logger.info("LocalBroadCastManager Received...")
@@ -166,7 +163,10 @@ class StreamingCore : Service(), AudioManager.OnAudioFocusChangeListener, Metada
                     Player.STATE_READY -> {
                         setPlayWhenReady(playWhenReady)
                     }
-                    else -> PlaybackStatus.IDLE
+                    else -> {
+                        pushEvent(FLUTTER_RADIO_PLAYER_STOPPED)
+                        PlaybackStatus.IDLE
+                    }
                 }
 
                 logger.info("onPlayerStateChanged: $playbackStatus")
@@ -184,15 +184,12 @@ class StreamingCore : Service(), AudioManager.OnAudioFocusChangeListener, Metada
         player.addMetadataOutput(this)
 
         createNotificationChannel()
-        notification = buildNotification(
-            initialTitle, subTitle, packageIntentName
-        )
-        startForeground(Companion.NOTIFICATION_ID, notification)
+        showNotification(initialTitle, subTitle, packageIntentName)
 
         return START_STICKY
     }
 
-    override fun onBind(intent: Intent?): IBinder? {
+    override fun onBind(intent: Intent?): IBinder {
         return iBinder
     }
 
@@ -221,36 +218,6 @@ class StreamingCore : Service(), AudioManager.OnAudioFocusChangeListener, Metada
         }
     }
 
-    override fun onAudioFocusChange(audioFocus: Int) {
-        when (audioFocus) {
-
-            AudioManager.AUDIOFOCUS_GAIN -> {
-                player.volume = 0.8f
-                play()
-                logger.info("AudioManager::AUDIO_FOCUS_GAIN")
-            }
-
-            AudioManager.AUDIOFOCUS_LOSS -> {
-                stop()
-                logger.info("AudioManager::AUDIO_FOCUS_LOSS")
-            }
-
-            AudioManager.AUDIOFOCUS_LOSS_TRANSIENT -> {
-                if (isPlaying()) {
-                    pause()
-                }
-                logger.info("AudioManager::AUDIO_FOCUS_LOSS_TRANSIENT")
-            }
-
-            AudioManager.AUDIOFOCUS_LOSS_TRANSIENT_CAN_DUCK -> {
-                if (isPlaying()) {
-                    player.volume = 0.1f
-                }
-                logger.info("AudioManager::AUDIO_FOCUS_LOSS_TRANSIENT_CAN_DUCK")
-            }
-        }
-    }
-
     override fun onMetadata(metadata: Metadata) {
         logger.info("onMetadata: $metadata")
 
@@ -265,12 +232,7 @@ class StreamingCore : Service(), AudioManager.OnAudioFocusChangeListener, Metada
             }
         }
 
-        notification = buildNotification(
-            getSongArtist(currentSong), getSongTitle(currentSong), packageIntentName
-        )?.apply {
-            notify(this)
-        }
-
+        showNotification(getSongArtist(currentSong), getSongTitle(currentSong), packageIntentName)
         localBroadcastManager.sendBroadcast(broadcastMetaDataIntent.putExtra("meta_data", currentSong))
     }
 
@@ -279,6 +241,8 @@ class StreamingCore : Service(), AudioManager.OnAudioFocusChangeListener, Metada
      */
     private fun kill() {
         telephonyManager?.listen(phoneStateListener, PhoneStateListener.LISTEN_NONE)
+        playerManager?.setPlayer(null)
+
         if (player != null) {
             player?.stop()
             if (playerEvents != null) {
@@ -286,7 +250,6 @@ class StreamingCore : Service(), AudioManager.OnAudioFocusChangeListener, Metada
             }
             player?.release()
         }
-        removeAudioFocus()
         try {
             if (becomingNoisyReceiver != null) {
                 unregisterReceiver(becomingNoisyReceiver)
@@ -295,33 +258,6 @@ class StreamingCore : Service(), AudioManager.OnAudioFocusChangeListener, Metada
             e.printStackTrace()
         }
         stopForeground(true)
-    }
-
-    /**
-     * Handle audio focus
-     */
-    private fun requestAudioFocus() {
-
-        val mAudioManager = getSystemService(Context.AUDIO_SERVICE) as AudioManager
-
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            audioFocusRequest?.let {
-                mAudioManager.requestAudioFocus(it)
-            }
-        } else {
-            mAudioManager.requestAudioFocus(context,
-                    AudioManager.STREAM_MUSIC,
-                    AudioManager.AUDIOFOCUS_GAIN)
-        }
-    }
-
-    private fun removeAudioFocus() {
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            val mAudioManager = getSystemService(Context.AUDIO_SERVICE) as AudioManager
-            audioFocusRequest?.let {
-                mAudioManager.abandonAudioFocusRequest(it)
-            }
-        }
     }
 
     /**
@@ -338,12 +274,13 @@ class StreamingCore : Service(), AudioManager.OnAudioFocusChangeListener, Metada
     private fun buildMediaSource(dataSourceFactory: DefaultDataSourceFactory, streamUrl: String): MediaSource {
 
         val uri = Uri.parse(streamUrl)
+        val mediaItem = MediaItem.fromUri(uri)
 
         return when (val type = Util.inferContentType(uri)) {
-            C.TYPE_DASH -> DashMediaSource.Factory(dataSourceFactory).createMediaSource(uri)
-            C.TYPE_SS -> SsMediaSource.Factory(dataSourceFactory).createMediaSource(uri)
-            C.TYPE_HLS -> HlsMediaSource.Factory(dataSourceFactory).createMediaSource(uri)
-            C.TYPE_OTHER -> ProgressiveMediaSource.Factory(dataSourceFactory).createMediaSource(uri)
+            C.TYPE_DASH -> DashMediaSource.Factory(dataSourceFactory).createMediaSource(mediaItem)
+            C.TYPE_SS -> SsMediaSource.Factory(dataSourceFactory).createMediaSource(mediaItem)
+            C.TYPE_HLS -> HlsMediaSource.Factory(dataSourceFactory).createMediaSource(mediaItem)
+            C.TYPE_OTHER -> ProgressiveMediaSource.Factory(dataSourceFactory).createMediaSource(mediaItem)
             else -> {
                 throw IllegalStateException("Unsupported type: $type")
             }
@@ -375,36 +312,90 @@ class StreamingCore : Service(), AudioManager.OnAudioFocusChangeListener, Metada
         }
     }
 
-    private fun buildNotification(
-            title: String,
-            description: String,
-            packageIntentName: String): Notification?
-    {
+    private fun showNotification(title: String, description: String, packageIntentName: String) {
         var notificationIntent: Intent? = null
 
         if (packageIntentName.isNotEmpty() && packageName.isNotEmpty()) {
             notificationIntent = packageManager.getLaunchIntentForPackage(packageName)
             notificationIntent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP)
         }
-        val pendingIntent: PendingIntent = PendingIntent.getActivity(this, 0, notificationIntent, 0)
+        val pendingIntent = PendingIntent.getActivity(this, 0, notificationIntent, 0)
 
-        val notificationBuilder = NotificationCompat.Builder(this, Companion.CHANNEL_ID)
-                .setVisibility(NotificationCompat.VISIBILITY_PUBLIC)
-                .setPriority(NotificationCompat.PRIORITY_LOW)
-                .setWhen(System.currentTimeMillis())
-                .setColorized(true)
-                .setContentText(description)
-                .setContentTitle(title)
-                .setAutoCancel(true)
-                .setSmallIcon(android.R.drawable.stat_sys_headset)
-                .setContentIntent(pendingIntent)
-        return notificationBuilder.build()
+        playerManager = PlayerNotificationManager.createWithNotificationChannel(
+            this,
+            "Channel_id",
+            me.sithiramunasinghe.flutter.flutter_radio_player.R.string.channel_name,
+            me.sithiramunasinghe.flutter.flutter_radio_player.R.string.channel_description,
+            Companion.NOTIFICATION_ID,
+            object : PlayerNotificationManager.MediaDescriptionAdapter {
+                override fun getCurrentContentTitle(player: Player): CharSequence {
+                    return title
+                }
+
+                override fun createCurrentContentIntent(player: Player): PendingIntent? {
+                    return pendingIntent
+                }
+
+                override fun getCurrentContentText(player: Player): CharSequence {
+                    return description
+                }
+
+                override fun getCurrentLargeIcon(
+                    player: Player,
+                    callback: PlayerNotificationManager.BitmapCallback
+                ): Bitmap? {
+                    return null
+                }
+            },
+            object : PlayerNotificationManager.NotificationListener {
+                override fun onNotificationPosted(
+                    notificationId: Int,
+                    notification: Notification,
+                    ongoing: Boolean
+                ) {
+                    startForeground(Companion.NOTIFICATION_ID, notification)
+                }
+
+                override fun onNotificationCancelled(
+                    notificationId: Int,
+                    dismissedByUser: Boolean
+                ) {
+                    kill()
+                }
+            }
+        ).apply {
+            setUseStopAction(false)
+            setUseNavigationActionsInCompactView(false)
+            setUseNavigationActions(false)
+            setPlayer(player)
+        }
     }
 
-    private fun notify(notification: Notification) {
-        val notificationManager = NotificationManagerCompat.from(this)
-        notificationManager.notify(Companion.NOTIFICATION_ID, notification)
-    }
+//    private fun buildNotification(
+//            title: String,
+//            description: String,
+//            packageIntentName: String): Notification?
+//    {
+//        var notificationIntent: Intent? = null
+//
+//        if (packageIntentName.isNotEmpty() && packageName.isNotEmpty()) {
+//            notificationIntent = packageManager.getLaunchIntentForPackage(packageName)
+//            notificationIntent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP)
+//        }
+//        val pendingIntent = PendingIntent.getActivity(this, 0, notificationIntent, 0)
+//        val notificationBuilder = NotificationCompat.Builder(this, Companion.CHANNEL_ID)
+//                .setVisibility(NotificationCompat.VISIBILITY_PUBLIC)
+//                .setPriority(NotificationCompat.PRIORITY_LOW)
+//                .setWhen(System.currentTimeMillis())
+//                .setStyle(androidx.media.app.NotificationCompat.MediaStyle())
+//                .setColorized(true)
+//                .setContentText(description)
+//                .setContentTitle(title)
+//                .setAutoCancel(true)
+//                .setSmallIcon(android.R.drawable.stat_sys_headset)
+//                .setContentIntent(pendingIntent)
+//        return notificationBuilder.build()
+//    }
 
     /**
      * Split metadata
