@@ -4,13 +4,13 @@ import 'package:flutter/services.dart';
 
 class FlutterRadioPlayer {
   static const MethodChannel _channel =
-  const MethodChannel('flutter_radio_player');
+      const MethodChannel('flutter_radio_player');
 
   static const EventChannel _eventChannel =
-  const EventChannel("flutter_radio_player_stream");
+      const EventChannel("flutter_radio_player_stream");
 
   static const EventChannel _eventChannelMetaData =
-  const EventChannel("flutter_radio_player_meta_stream");
+      const EventChannel("flutter_radio_player_meta_stream");
 
   // constants to support event channel
   static const flutter_radio_stopped = "flutter_radio_stopped";
@@ -19,34 +19,34 @@ class FlutterRadioPlayer {
   static const flutter_radio_error = "flutter_radio_error";
   static const flutter_radio_loading = "flutter_radio_loading";
 
-  static Stream<String> _isPlayingStream;
-  static Stream<String> _metaDataStream;
+  static Stream<String>? _isPlayingStream;
+  static Stream<String>? _metaDataStream;
 
   Future<void> init(String initialTitle, String subTitle, String streamURL,
-      String playWhenReady) async {
+      bool playWhenReady) async {
     return await _channel.invokeMethod("initService", {
       "initialTitle": initialTitle,
       "subTitle": subTitle,
       "streamURL": streamURL,
-      "playWhenReady": playWhenReady
+      "playWhenReady": playWhenReady ? "true" : "false"
     });
   }
 
   Future<bool> play() async {
-    return await _channel.invokeMethod("play");
+    return await _channel.invokeMethod("play") ?? false;
   }
 
   Future<bool> pause() async {
-    return await _channel.invokeMethod("pause");
+    return await _channel.invokeMethod("pause") ?? false;
   }
 
   Future<bool> playOrPause() async {
     print("Invoking platform method: playOrPause");
-    return await _channel.invokeMethod("playOrPause");
+    return await _channel.invokeMethod("playOrPause") ?? false;
   }
 
   Future<bool> stop() async {
-    return await _channel.invokeMethod("stop");
+    return await _channel.invokeMethod("stop") ?? false;
   }
 
   Future<bool> isPlaying() async {
@@ -59,10 +59,8 @@ class FlutterRadioPlayer {
   }
 
   Future<void> setUrl(String streamUrl, String playWhenReady) async {
-    await _channel.invokeMethod("setUrl", {
-      "playWhenReady": playWhenReady,
-      "streamUrl": streamUrl
-    });
+    await _channel.invokeMethod(
+        "setUrl", {"playWhenReady": playWhenReady, "streamUrl": streamUrl});
   }
 
   Future<String> currentSongTitle() async {
@@ -70,7 +68,7 @@ class FlutterRadioPlayer {
   }
 
   /// Get the player stream.
-  Stream<String> get isPlayingStream {
+  Stream<String>? get isPlayingStream {
     if (_isPlayingStream == null) {
       _isPlayingStream =
           _eventChannel.receiveBroadcastStream().map<String>((value) => value);
@@ -78,10 +76,11 @@ class FlutterRadioPlayer {
     return _isPlayingStream;
   }
 
-  Stream<String> get metaDataStream {
+  Stream<String>? get metaDataStream {
     if (_metaDataStream == null) {
-      _metaDataStream =
-          _eventChannelMetaData.receiveBroadcastStream().map<String>((value) => value);
+      _metaDataStream = _eventChannelMetaData
+          .receiveBroadcastStream()
+          .map<String>((value) => value);
     }
 
     return _metaDataStream;
